Drop direct scopeguard dependency on Windows
